+++ NEWS.md
@@ -3,10 +3,11 @@
 ## v0.3.0
 
 * [#143](https://github.com/brodieG/diffobj/issues/143): Add `ses_dat` to
-  provide a more computable version of `ses` (h/t @hadley)
+  provide a more computable version of `ses` (h/t @hadley).
 * [#144](https://github.com/brodieG/diffobj/issues/144): Re-encode strings to
   UTF-8 prior to comparison to avoid spurious encoding-only differences (h/t
-  @hadley)
+  @hadley).
+* Banish ghosts of `stringsAsFactors`.
 
 ## v0.2.4